A lot of changes

+++ models/truck.py
@@ -1,5 +1,5 @@
 class Truck:
-    def __init__(self, id: int, name: str, capacity: float, max_range: int):
+    def __init__(self, id: int, name: str, capacity: int, max_range: int):
         self._id = id
         self._name = name
         self._capacity = capacity
